testing/mm: Update after the recent changes in nuttx
The following commit changed the typo of SIZEOF_MM_ALLOCNODE
from int to size_t.
```
commit 39eaeefb78f36724adbdc47400d6f60372b68344
Author: Jiuzhu Dong <dongjiuzhu1@xiaomi.com>
Date:   Thu Jan 27 12:17:20 2022 +0800

    mm/mm_heap: remove the unnecessary check for memory node size


```

+++ testing/mm/mm_main.c
@@ -351,7 +351,8 @@ static void *realloc_boundary_malloc(int *nodesize)
     }
   else
     {
-      printf("malloc failed, size=%d\n", (size << 1) - SIZEOF_MM_ALLOCNODE);
+      printf("malloc failed, size=%zu\n",
+             (size_t)((size << 1) - SIZEOF_MM_ALLOCNODE));
     }
 
   return ptr;
@@ -374,7 +375,7 @@ static void realloc_boundary(void)
   DEBUGASSERT(sizeof(dq_entry_t) <= (MM_MIN_CHUNK - SIZEOF_MM_ALLOCNODE));
 
   printf("memory realloc_boundary test start.\n");
-  printf("MM_MIN_CHUNK=%d, SIZEOF_MM_ALLOCNODE=%d\n",
+  printf("MM_MIN_CHUNK=%d, SIZEOF_MM_ALLOCNODE=%zu\n",
          MM_MIN_CHUNK, SIZEOF_MM_ALLOCNODE);
 
   /* Malloc memory until the memeory ran out */
